#1, #2 : add new variable to GameSpace, which will check if is countdown to fix bug with keyevents

+++ CardsGame/Models/GameCounters.cs
@@ -21,7 +21,7 @@ namespace CardsGame.Models
         /// <summary>
         /// Reached points
         /// </summary>
-        public int TotalPoints { get; private set; } = 0;
+        public long TotalPoints { get; private set; } = 0;
 
         /// <summary>
         /// Variable to hold the best streak
